final/final-django/requirements.txt
absl-py==0.15.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
flatbuffers==1.12
fonttools==4.51.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
jax==0.4.28
jaxlib==0.4.28
keras==3.3.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.19.5
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scipy==1.13.0
six==1.15.0
sounddevice==0.4.6
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
typing-extensions==3.7.4.3
tzdata==2023.3
urllib3==2.0.4
Werkzeug==3.0.3
wrapt==1.12.1
zipp==3.18.1
